퍼센트화 수정 및 전반적인 오류 찾기

+++ src/main/java/com/koreait/exam/tmrteamproject/repository/DataSetRepository.java
@@ -94,7 +94,7 @@ public interface DataSetRepository extends JpaRepository<DataSet, Long> {
     // 퍼센트용 계산 구하기
     // 평균 매출액
     @Query("SELECT AVG(d.monthlySalesAmount) FROM DataSet d WHERE d.baseYearQuarterCode = :quarter AND d.adminDongCode = :adminDongCode")
-    Double findAvgSalesByQuarter(@Param("quarter") String quarter, String adminDongCode);
+    Long findAvgSalesByQuarter(@Param("quarter") String quarter, String adminDongCode);
 
     // 분기별 행정동별 점포수 합
     @Query("SELECT SUM(d.storeCount) FROM DataSet d WHERE d.baseYearQuarterCode = :quarter AND d.adminDongCode = :adminDongCode")
@@ -102,5 +102,5 @@ public interface DataSetRepository extends JpaRepository<DataSet, Long> {
 
     //분기별 행정동별 평균 유동인구
     @Query("SELECT AVG(d.totalFloatingPopulation) FROM DataSet d WHERE d.baseYearQuarterCode = :quarter AND d.adminDongCode = :adminDongCode")
-    Double findAvgFloatingByQuarter(@Param("quarter") String quarter, String adminDongCode);
+    Long findAvgFloatingByQuarter(@Param("quarter") String quarter, String adminDongCode);
 }
